+++ xtrack/custom_setter/__init__.py
@@ -1,0 +1,1 @@
+from .custom_setter import CustomSetter
